various integer size fixen. Some of these are kind of blind narrowing of int types without checking for potential overflow. It doesn't *seem* like any of these will overflow / wrap-around, but it's not impossible.

+++ libraries/Boost/boost_1_54_0/boost/date_time/posix_time/conversion.hpp
@@ -27,7 +27,7 @@ namespace posix_time {
   ptime from_time_t(std::time_t t)
   {
     ptime start(gregorian::date(1970,1,1));
-    return start + seconds(static_cast<long>(t));
+    return start + seconds(static_cast<int>(t));
   }
 
   //! Convert a time to a tm structure truncating any fractional seconds
